+++ elasticsearch-20170613/src/Models/ListInstanceResponseBody/result/zoneInfos.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Elasticsearch\V20170613\Models\ListInstanceResponseBody\result;
+
+use AlibabaCloud\Dara\Model;
+
+class zoneInfos extends Model
+{
+    /**
+     * @var string
+     */
+    public $status;
+
+    /**
+     * @var string
+     */
+    public $zoneId;
+    protected $_name = [
+        'status' => 'status',
+        'zoneId' => 'zoneId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->status) {
+            $res['status'] = $this->status;
+        }
+
+        if (null !== $this->zoneId) {
+            $res['zoneId'] = $this->zoneId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['status'])) {
+            $model->status = $map['status'];
+        }
+
+        if (isset($map['zoneId'])) {
+            $model->zoneId = $map['zoneId'];
+        }
+
+        return $model;
+    }
+}